requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
tqdm>=4.64.0
pillow>=9.0.0
streamlit>=1.18.0
qrcode>=8.1